implemented: layout with viewport size
