+++ src/Domain/Layout/Layout.cpp
@@ -11,122 +11,30 @@ using namespace VGG::Layout::Internal::Rule;
 
 constexpr auto FLIP_Y_FACTOR = -1;
 
-namespace
-{
-using LengthTypes = Layout::Internal::Rule::Length::Types;
-unit toLibUnit(LengthTypes type)
-{
-  switch (type)
-  {
-    case LengthTypes::px:
-      return unit_point;
-    case LengthTypes::percent:
-      return unit_percent;
-    case LengthTypes::fit_content:
-      return unit_auto;
-    default:
-      return unit_undefine;
-  }
-}
-
-position toLibPosition(Position::Types type)
-{
-  switch (type)
-  {
-    case Position::Types::Relative:
-      return position_relative;
-    case Position::Types::Absolute:
-      return position_absolute;
-
-    default: // todo
-      return position_absolute;
-  }
-}
-
-direction toLibDirecion(FlexboxLayout::Direction type)
-{
-  switch (type)
-  {
-    case FlexboxLayout::Direction::Horizontal:
-      return direction_row;
-
-    default:
-      return direction_column;
-  }
-}
-
-justify_content toLibJustifyContent(FlexboxLayout::JustifyContent type)
-{
-  switch (type)
-  {
-    case FlexboxLayout::JustifyContent::Start:
-      return justify_content_flex_start;
-    case FlexboxLayout::JustifyContent::Center:
-      return justify_content_center;
-    case FlexboxLayout::JustifyContent::End:
-      return justify_content_flex_end;
-    case FlexboxLayout::JustifyContent::SpaceBetween:
-      return justify_content_space_between;
-    case FlexboxLayout::JustifyContent::SpaceAround:
-      return justify_content_space_around;
-    case FlexboxLayout::JustifyContent::SpaceEvenly:
-      return justify_content_space_evenly;
-  }
-}
-
-align_content toLibAlignContent(AlignStyle type)
-{
-  switch (type)
-  {
-    case AlignStyle::Start:
-      return align_content_flex_start;
-    case AlignStyle::Center:
-      return align_content_center;
-    case AlignStyle::End:
-      return align_content_flex_end;
-  }
-}
-
-align_items toLibAlignItem(AlignStyle type)
-{
-  switch (type)
-  {
-    case AlignStyle::Start:
-      return align_items_flex_start;
-    case AlignStyle::Center:
-      return align_items_center;
-    case AlignStyle::End:
-      return align_items_flex_end;
-  }
-}
-
-wrap toLibWrap(FlexboxLayout::Wrap type)
-{
-  switch (type)
-  {
-    case FlexboxLayout::Wrap::NoWrap:
-      return wrap_no_wrap;
-    case FlexboxLayout::Wrap::Wrap:
-      return wrap_wrap;
-  }
-}
-
-} // namespace
-
 void Layout::Layout::layout(Size size)
 {
-  auto view = layoutTree();
+  auto root = layoutTree();
 
   if (size != m_size)
   {
     m_size = size;
 
-    auto frame = view->frame();
-    frame.size = m_size;
-    view->setFrame(frame);
+    // initial config
+    configureAutoLayout(root);
 
-    configureView(view);
-    view->applyLayout();
+    // udpate Frames's frame
+    auto frame = root->frame();
+    frame.size = m_size;
+    root->setFrame(frame);
+    for (auto& subview : root->children())
+    {
+      auto frame = subview->frame();
+      frame.size = m_size;
+      subview->setFrame(frame);
+    }
+
+    //
+    root->layoutIfNeeded();
   }
 }
 
@@ -247,7 +155,7 @@ void Layout::Layout::collectRules(const nlohmann::json& json)
   }
 }
 
-void Layout::Layout::configureView(std::shared_ptr<LayoutView> view)
+void Layout::Layout::configureAutoLayout(std::shared_ptr<LayoutView> view)
 {
   std::shared_ptr<VGG::Layout::Internal::Rule::Rule> rule;
 
@@ -261,118 +169,15 @@ void Layout::Layout::configureView(std::shared_ptr<LayoutView> view)
     }
   }
 
-  auto autoLayout = view->resetAutoLayout();
+  auto autoLayout = view->createAutoLayout();
   if (rule)
   {
-    autoLayout->isEnabled = true;
     autoLayout->rule = rule;
-
-    autoLayout->flexNode = nullptr;
-    autoLayout->gridNode = nullptr;
-
-    configureView(view, rule);
+    view->configureAutoLayout();
   }
 
   for (auto& child : view->children())
   {
-    configureView(child);
-  }
-}
-
-void Layout::Layout::configureView(std::shared_ptr<LayoutView> view,
-                                   std::shared_ptr<VGG::Layout::Internal::Rule::Rule> rule)
-{
-  if (const auto detail = std::get_if<FlexboxLayout>(&rule->layout))
-  {
-    DEBUG("Layout::configureView, flex layout, view[%p, %s]", view.get(), view->path().c_str());
-
-    auto node = view->createFlexboxNode();
-    configureNode(node, rule);
-
-    node->set_direction(toLibDirecion(detail->direction));
-    node->set_justify_content(toLibJustifyContent(detail->justify_content));
-    node->set_align_items(toLibAlignItem(detail->align_items));
-    node->set_align_content(toLibAlignContent(detail->align_content));
-    node->set_wrap(toLibWrap(detail->wrap));
-    node->set_gap(gap_row, detail->row_gap);
-    node->set_gap(gap_column, detail->column_gap);
-    node->set_padding(padding_top, detail->padding.top);
-    node->set_padding(padding_right, detail->padding.right);
-    node->set_padding(padding_bottom, detail->padding.bottom);
-    node->set_padding(padding_left, detail->padding.left);
-  }
-  else if (const auto detail = std::get_if<GridLayout>(&rule->layout))
-  {
-    DEBUG("Layout::configureView, grid layout, view[%p, %s]", view.get(), view->path().c_str());
-
-    // todo
-    // auto node = new grid_layout;
-    // autoLayout->gridNode.reset(node);
-  }
-
-  if (const auto detail = std::get_if<FlexboxItem>(&rule->item_in_layout))
-  {
-    DEBUG("Layout::configureView, flex item , view[%p, %s]", view.get(), view->path().c_str());
-
-    auto node = view->autoLayout()->flexNode;
-    if (!node)
-    {
-      node = view->createFlexboxNode();
-    }
-    configureNode(node, rule);
-
-    node->set_position(toLibPosition(detail->position.value));
-    node->set_grow(detail->flex_grow);
-
-    if (detail->top.has_value())
-    {
-      node->set_ltrb(ltrb_top, detail->top->value);
-    }
-    if (detail->right.has_value())
-    {
-      node->set_ltrb(ltrb_right, detail->right->value);
-    }
-    if (detail->bottom.has_value())
-    {
-      node->set_ltrb(ltrb_bottom, detail->bottom->value);
-    }
-    if (detail->left.has_value())
-    {
-      node->set_ltrb(ltrb_left, detail->left->value);
-    }
-  }
-  else if (const auto detail = std::get_if<GridItem>(&rule->item_in_layout))
-  {
-    DEBUG("Layout::configureView, grid item , view[%p, %s]", view.get(), view->path().c_str());
-    // todo
-  }
-}
-
-void Layout::Layout::configureNode(std::shared_ptr<flexbox_node> node,
-                                   std::shared_ptr<VGG::Layout::Internal::Rule::Rule> rule)
-{
-  node->set_width(toLibUnit(rule->width.value.types), rule->width.value.value);
-  if (rule->max_width.has_value())
-  {
-    node->set_max_width(toLibUnit(rule->max_width->value.types), rule->max_width->value.value);
-  }
-  if (rule->min_width.has_value())
-  {
-    node->set_min_width(toLibUnit(rule->min_width->value.types), rule->min_width->value.value);
-  }
-
-  node->set_height(toLibUnit(rule->height.value.types), rule->height.value.value);
-  if (rule->max_height.has_value())
-  {
-    node->set_max_height(toLibUnit(rule->max_height->value.types), rule->max_height->value.value);
-  }
-  if (rule->min_height.has_value())
-  {
-    node->set_min_height(toLibUnit(rule->min_height->value.types), rule->min_height->value.value);
-  }
-
-  if (rule->aspect_ratio.has_value())
-  {
-    // todo
+    configureAutoLayout(child);
   }
 }
