+++ test/domain/layout/layout_tests.cpp
@@ -58,7 +58,7 @@ TEST_F(VggLayoutTestSuite, Layout)
   EXPECT_TRUE(left_child_frame == expect_left_child_frame);
 
   auto right_child_frame = current_page->children()[1]->frame();
-  Layout::Rect expect_right_child_frame{ { 1100, 40 }, { 200, 300 } };
+  Layout::Rect expect_right_child_frame{ { 1100, 40 }, { 200, 250 } };
   EXPECT_TRUE(right_child_frame == expect_right_child_frame);
 
   // todo, check json model
